+++ lwjgl3/src/main/java/dev/mariorez/lwjgl3/Lwjgl3Launcher.java
@@ -25,7 +25,7 @@ public class Lwjgl3Launcher {
         //// If you remove the above line and set Vsync to false, you can get unlimited FPS, which can be
         //// useful for testing performance, but can also be very stressful to some hardware.
         //// You may also need to configure GPU drivers to fully disable Vsync; this can cause screen tearing.
-        configuration.setWindowedMode(WINDOW_WIDTH, WINDOW_HEIGHT);
+        configuration.setWindowedMode((int) WINDOW_WIDTH, (int) WINDOW_HEIGHT);
         configuration.setWindowIcon("libgdx128.png", "libgdx64.png", "libgdx32.png", "libgdx16.png");
         return configuration;
     }
